move commit search to internal/search/commits
