+++ internal/search/commits/commits.go
@@ -1,4 +1,4 @@
-package run
+package commits
 
 import (
 	"bytes"
