+++ internal/search/run/textsearch_test.go
@@ -82,7 +82,7 @@ func TestSearchFilesInRepos(t *testing.T) {
 	repoRevs := makeRepositoryRevisions("foo/one", "foo/two", "foo/empty", "foo/cloning", "foo/missing", "foo/missing-database", "foo/timedout", "foo/no-rev")
 	args := &search.TextParameters{
 		PatternInfo: &search.TextPatternInfo{
-			FileMatchLimit: defaultMaxSearchResults,
+			FileMatchLimit: search.DefaultMaxSearchResults,
 			Pattern:        "foo",
 		},
 		RepoPromise:  (&search.RepoPromise{}).Resolve(repoRevs),
@@ -112,7 +112,7 @@ func TestSearchFilesInRepos(t *testing.T) {
 	// that should be checked earlier.
 	args = &search.TextParameters{
 		PatternInfo: &search.TextPatternInfo{
-			FileMatchLimit: defaultMaxSearchResults,
+			FileMatchLimit: search.DefaultMaxSearchResults,
 			Pattern:        "foo",
 		},
 		RepoPromise:  (&search.RepoPromise{}).Resolve(makeRepositoryRevisions("foo/no-rev@dev")),
@@ -169,7 +169,7 @@ func TestSearchFilesInReposStream(t *testing.T) {
 	}
 	args := &search.TextParameters{
 		PatternInfo: &search.TextPatternInfo{
-			FileMatchLimit: defaultMaxSearchResults,
+			FileMatchLimit: search.DefaultMaxSearchResults,
 			Pattern:        "foo",
 		},
 		RepoPromise:  (&search.RepoPromise{}).Resolve(makeRepositoryRevisions("foo/one", "foo/two", "foo/three")),
@@ -235,7 +235,7 @@ func TestSearchFilesInRepos_multipleRevsPerRepo(t *testing.T) {
 	}
 	args := &search.TextParameters{
 		PatternInfo: &search.TextPatternInfo{
-			FileMatchLimit: defaultMaxSearchResults,
+			FileMatchLimit: search.DefaultMaxSearchResults,
 			Pattern:        "foo",
 		},
 		RepoPromise:  (&search.RepoPromise{}).Resolve(makeRepositoryRevisions("foo@master:mybranch:*refs/heads/")),
@@ -283,7 +283,7 @@ func TestRepoShouldBeSearched(t *testing.T) {
 	}
 	defer func() { searcher.MockSearch = nil }()
 	info := &search.TextPatternInfo{
-		FileMatchLimit:               defaultMaxSearchResults,
+		FileMatchLimit:               search.DefaultMaxSearchResults,
 		Pattern:                      "foo",
 		FilePatternsReposMustInclude: []string{"main"},
 	}
