+++ internal/search/search.go
@@ -1,0 +1,3 @@
+package search
+
+const DefaultMaxSearchResults = 30
